chamcong/service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
facenet-pytorch==2.6.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0
python-multipart==0.0.9
pyyaml==6.0.1
requests==2.32.5
